+++ app/react/V2/Components/Metadata/MetadataDisplay.tsx
@@ -20,19 +20,8 @@ import {
   Image,
   Media,
 } from './Components/index.js';
-import { BaseMetadataProperty, MetadataProperty } from '#V2/formatters/types.js';
-import {
-  formatRelationshipProperty,
-  formatSimpleProperty,
-  formatDateProperty,
-  formatGeolocationProperty,
-  formatLinkProperty,
-  formatMediaProperty,
-  formatImageProperty,
-  formatMetadataFields,
-  formatSelectProperty,
-} from '#V2/formatters/index.js';
-import { resolvePropertyType } from '#V2/formatters/metadata/resolvePropertyMetadataValues.js';
+import { MetadataProperty } from '#V2/formatters/types.js';
+import { useFormatMetadata } from './hooks/useFormatMetadata.js';
 
 type MetadataDisplayProps = {
   entity: Entity;
@@ -40,69 +29,10 @@ type MetadataDisplayProps = {
 
 const MetadataDisplay = ({ entity }: MetadataDisplayProps) => {
   const templates = useAtomValue(templatesAtom);
-  const entityTemplate = useMemo(
-    () => templates.find(template => template._id === entity.template),
-    [entity.template, templates]
-  );
 
-  const metadataFields: BaseMetadataProperty[] = useMemo(
-    () => formatMetadataFields(entityTemplate, { groupGeolocationProperties: true }),
-    [entityTemplate]
-  );
-
-  const metadata: MetadataProperty[] = useMemo(
-    () =>
-      metadataFields
-        .map(field => {
-          const type = resolvePropertyType(field, entity.metadata);
-
-          if (
-            type === 'text' ||
-            type === 'generatedid' ||
-            type === 'numeric' ||
-            type === 'markdown'
-          ) {
-            return formatSimpleProperty(field, entity.metadata);
-          }
-
-          if (
-            type === 'date' ||
-            type === 'daterange' ||
-            type === 'multidate' ||
-            type === 'multidaterange'
-          ) {
-            return formatDateProperty(field, entity.metadata);
-          }
-
-          if (type === 'geolocation') {
-            return formatGeolocationProperty(field, entity, templates);
-          }
-
-          if (type === 'select' || type === 'multiselect') {
-            return formatSelectProperty(field, entity.metadata);
-          }
-
-          if (type === 'link') {
-            return formatLinkProperty(field, entity.metadata);
-          }
-
-          if (type === 'media') {
-            return formatMediaProperty(field, entity.metadata);
-          }
-
-          if (type === 'image' || type === 'preview') {
-            return formatImageProperty(field, entity.metadata, entityTemplate);
-          }
-
-          if (type === 'relationship') {
-            return formatRelationshipProperty(field, entity.metadata);
-          }
-
-          return null;
-        })
-        .filter(m => m) as MetadataProperty[],
-    [entity, metadataFields, entityTemplate, templates]
-  );
+  const { entityTemplate, metadata } = useFormatMetadata(entity, templates, {
+    groupGeolocationProperties: true,
+  });
 
   const renderMetadataFields = useCallback(
     (fields: MetadataProperty[]) => {
